# docs/swagger.yaml
basePath: /
definitions:
  domain.Account:
    properties:
      account_id:
        type: integer
      document_number:
        type: string
    type: object
  domain.Transaction:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      event_date:
        type: string
      operation_type_id:
        type: integer
      transaction_id:
        type: integer
    type: object
  http.createAccountRequest:
    properties:
      document_number:
        type: string
    type: object
  http.createTransactionRequest:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      operation_type_id:
        type: integer
    type: object
  http.errorResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  description: REST API for the Pismo code assessment.
  title: Pismo Assessment API
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Create an account
  /accounts/{accountId}:
    get:
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Get an account by ID
  /transactions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Create a transaction
swagger: "2.0"